+++ entity_linkage/normalization/ned_with_web_links/tests/Readme.md
@@ -1,0 +1,3 @@
+This paper load wikipedia data by calling sift package.<br>
+The input format is (_id, [(start, stop, target),...], text).<br>
+The evaluation metrics is P@1 in this paper.
